update all user ids and references to be long, reimplement chat screen and navigation with Bill's changes, and abstract sample chat partners data through repository

+++ app/src/main/java/com/swen549/touchanalytics/data/Feature.kt
@@ -1,7 +1,7 @@
 package com.swen549.touchanalytics.data
 
 data class Feature(
-    val userId: Int = 0,
+    val userId: Long = 0,
     val startX: Float = 0f,
     val stopX: Float = 0f,
     val startY: Float = 0f,
